ServerThread Sends list of Cards to Each Client
Player turns are now decided by a PlayerQueue
ServerThread sends list of each players cards to the respective client.
This is the new group 1.
Adjusted ClientThread's State 3 to accomadate if it is updating a
Client's list of cards.

Client State 3 now checks if it's message comes with a code at mess[0].
Other parts are moved to the next index